pyproject.toml
[tool.poetry]
name = "kruxapp"
version = "25.05.beta7"
description = "Krux Android app for learning about Bitcoin air-gapped transactions"
authors = ["Krux Team <contact@krux.org>"]
maintainers = ["Krux Team <contact@krux.org>"]
readme = "README.md"
license = "MIT"
homepage = "https://selfcustody.com/krux"
repository = "https://github.com/selfcustody/krux"
keywords = ["bitcoin", "kivy", "android", "air-gapped", "wallet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Kivy",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security :: Cryptography",
]
packages = [
    { include = "src" },
    { include = "mocks" },
    { include = "camerax_provider" },
]
include = [
    "buildozer.spec",
    "*.ttf",
    "data/*.png",
    "data/*.ico",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
kivy = "^2.3.0"
pyqrcode = "^1.2.1"
pyzbar = "^0.1.9"
gestures4kivy = "^0.1.4"
camera4kivy = "^0.3.3"
pycryptodome = "^3.20.0"
pillow = "^10.0.0"
numpy = "^1.24.0"
opencv-python = "^4.11.0.86"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
black = "^24.4.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
pre-commit = "^3.7.0"

[tool.poetry.group.build.dependencies]
buildozer = "^1.5.0"

[tool.poetry.scripts]
krux = "main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '''
/(
  \.pyi?|
  /src/|
  /mocks/|
  /camerax_provider/
)$
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
